removed vector from optimizer

+++ src/optimization/optimizer.hpp
@@ -11,6 +11,6 @@ namespace opt
 
 fin::Portfolio get_optimal_portfolio(fin::Asset *h_assets, int *port_assets,
                                       hlp::Date& d1, hlp::Date& d2,
-                                      size_t n, size_t nb_p = 10, size_t k = 20);
+                                      int n, int nb_p = 10, int k = 20);
 
 }
